refactor: move matrix_plotting under nilearn.plotting.matrix
