+++ nilearn/plotting/matrix/__init__.py
@@ -1,0 +1,16 @@
+from nilearn.plotting.matrix.matrix_plotting import (
+    plot_contrast_matrix,
+    plot_design_matrix,
+    plot_design_matrix_correlation,
+    plot_event,
+    plot_matrix,
+)
+
+
+__all__ = [
+    "plot_contrast_matrix",
+    "plot_design_matrix",
+    "plot_design_matrix_correlation",
+    "plot_event",
+    "plot_matrix",
+]
